# Bloggify-swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Bloggify",
    "contact": {}
  },
  "host": "localhost:5000",
  "basePath": "/api/user",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/new-blog": {
      "post": {
        "summary": "Add New Blog",
        "tags": [
          "Blogs"
        ],
        "operationId": "AddNewBlog",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "category_id",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "title",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "content",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "status",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AddNewBlog"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "status": 1,
                "message": "Blog Published Successfully.",
                "data": {
                  "category_id": "6197cd59e28a5eb64301b2c8",
                  "image": "https://bloggify.s3.amazonaws.com/90dee1de-9a03-493a-90b2-c37cd4079306.jpeg",
                  "image_alt": "90dee1de-9a03-493a-90b2-c37cd4079306",
                  "title": "Why should you become a full stack developer?",
                  "author_id": "6199039f2b88fabc684ab32f",
                  "content": "Let’s get this cleared first. Here are the top 3 reasons why you should consider becoming a full stack developer -\n\n1. Building unique proprietary software products\nEvery company/developer follows their own tech stack while building a product (as every product has its own requirements and the best set of technology is chosen accordingly).\n\nThe use of such diverse technologies makes the code base of the project complex and this eventually creates a unique code for a particular product. Companies adopting full stack development will have this edge.",
                  "status": "published",
                  "date": "2021-11-20T14:23:37.117Z",
                  "_id": "619904e92b88fabc684ab332",
                  "createdAt": "2021-11-20T14:23:37.130Z",
                  "updatedAt": "2021-11-20T14:23:37.130Z"
                }
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "1073"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"431-cPa7rDWEo/l4+lvY3tKNLEiY9HQ\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 20 Nov 2021 14:23:37 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
    "/user/update-blog": {
      "put": {
        "summary": "Update Blog",
        "tags": [
          "Blogs"
        ],
        "operationId": "UpdateBlog",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "image",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "image_alt",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "content",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "status",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/UpdateBlog-UnAuthorized1"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "status": 0,
                "message": "You are not Authorized!",
                "data": null
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "60"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"3c-SOUeNEqcTolR0CzbGH0NayJlX38\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 20 Nov 2021 14:25:50 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdateBlog-Authorized"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "status": 1,
                "message": "Blog Updated Successfully.",
                "data": {
                  "_id": "6197cf099d6a0fa882cb0e75",
                  "category_id": "6197cd59e28a5eb64301b2c8",
                  "image": "https://bloggify.s3.amazonaws.com/90dee1de-9a03-493a-90b2-c37cd4079306.jpeg",
                  "image_alt": "90dee1de-9a03-493a-90b2-c37cd4079306",
                  "title": "Why should you become a full stack developer?",
                  "author_id": "61977e099fe7b649dae5e0c0",
                  "content": "This diverse working knowledge gives developers the flexibility to tailor the tools needed for the project as per the project needs and create a state of the art products; since every technology has something different to offer - broadly in terms of syntaxes, compiling time (performance), compatibility, code reusability.\n\nOne fine example is that web developers with front-end knowledge can easily convert animations and graphics offered by software like Adobe Photoshop directly to front-end code. Thereby reducing the dependency on such third-party proprietary software. Limiting knowledge of basic styling tools will bind the developer to use such third-party software and thus burdening the project’s budget. Similar advantages apply to back-end scenarios. And thus knowing several essential tools will open up innovative approaches to projects.",
                  "status": "published",
                  "date": "2021-11-19T16:21:29.073Z",
                  "createdAt": "2021-11-19T16:21:29.109Z",
                  "updatedAt": "2021-11-20T14:26:12.847Z"
                }
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "1367"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"557-k8bmVT2CROGIMe/xzHXIekQoAR0\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 20 Nov 2021 14:26:12 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
    "/user/delete-blog": {
      "delete": {
        "summary": "Delete Blog",
        "tags": [
          "Blogs"
        ],
        "operationId": "DeleteBlog",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/DeleteBlog-NotFound"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "status": 0,
                "message": "Blog Not Found!",
                "data": null
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "52"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"34-QLcH6v3QA+qtxwzXD+CLLEkhyJ8\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 20 Nov 2021 14:27:05 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/DeleteBlog-NotAuthorized1"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "status": 0,
                "message": "You are not Authorized!",
                "data": null
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "60"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"3c-SOUeNEqcTolR0CzbGH0NayJlX38\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 20 Nov 2021 14:27:29 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeleteBlog-Success"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "status": 1,
                "message": "Blog Deleted Successfully.",
                "data": {
                  "_id": "619904e92b88fabc684ab332",
                  "category_id": "6197cd59e28a5eb64301b2c8",
                  "image": "https://bloggify.s3.amazonaws.com/90dee1de-9a03-493a-90b2-c37cd4079306.jpeg",
                  "image_alt": "90dee1de-9a03-493a-90b2-c37cd4079306",
                  "title": "Why should you become a full stack developer?",
                  "author_id": "6199039f2b88fabc684ab32f",
                  "content": "Let’s get this cleared first. Here are the top 3 reasons why you should consider becoming a full stack developer -\n\n1. Building unique proprietary software products\nEvery company/developer follows their own tech stack while building a product (as every product has its own requirements and the best set of technology is chosen accordingly).\n\nThe use of such diverse technologies makes the code base of the project complex and this eventually creates a unique code for a particular product. Companies adopting full stack development will have this edge.",
                  "status": "published",
                  "date": "2021-11-20T14:23:37.117Z",
                  "createdAt": "2021-11-20T14:23:37.130Z",
                  "updatedAt": "2021-11-20T14:23:37.130Z"
                }
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "1071"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"42f-4BeYJoTXWBq76rrtKz23sNvVdUM\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 20 Nov 2021 14:28:11 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
...
    "/user/find-user": {
      "get": {
        "summary": "Search User",
        "tags": [
          "Users"
        ],
        "operationId": "SearchUser",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SearchUser-Success"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "status": 1,
                "message": "User Found Success.",
                "data": {
                  "email": "hkanani191@gmail.com",
                  "usernmae": "harikanani",
                  "full_name": "hari kanani"
                }
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "134"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"86-3vXlxw2a/t8jL1o0mK9YoG9eJo0\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 20 Nov 2021 14:17:20 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AddNewBlog": {
      "title": "AddNewBlog",
      "example": {
        "status": 1,
        "message": "Blog Published Successfully.",
        "data": {
          "category_id": "6197cd59e28a5eb64301b2c8",
          "image": "https://bloggify.s3.amazonaws.com/90dee1de-9a03-493a-90b2-c37cd4079306.jpeg",
          "image_alt": "90dee1de-9a03-493a-90b2-c37cd4079306",
          "title": "Why should you become a full stack developer?",
          "author_id": "6199039f2b88fabc684ab32f",
          "content": "Let’s get this cleared first. Here are the top 3 reasons why you should consider becoming a full stack developer -\n\n1. Building unique proprietary software products\nEvery company/developer follows their own tech stack while building a product (as every product has its own requirements and the best set of technology is chosen accordingly).\n\nThe use of such diverse technologies makes the code base of the project complex and this eventually creates a unique code for a particular product. Companies adopting full stack development will have this edge.",
          "status": "published",
          "date": "2021-11-20T14:23:37.117Z",
          "_id": "619904e92b88fabc684ab332",
          "createdAt": "2021-11-20T14:23:37.130Z",
          "updatedAt": "2021-11-20T14:23:37.130Z"
        }
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/Data2"
        }
      },
      "required": [
        "status",
        "message",
        "data"
      ]
    },
    "Data2": {
      "title": "Data2",
      "example": {
        "category_id": "6197cd59e28a5eb64301b2c8",
        "image": "https://bloggify.s3.amazonaws.com/90dee1de-9a03-493a-90b2-c37cd4079306.jpeg",
        "image_alt": "90dee1de-9a03-493a-90b2-c37cd4079306",
        "title": "Why should you become a full stack developer?",
        "author_id": "6199039f2b88fabc684ab32f",
        "content": "Let’s get this cleared first. Here are the top 3 reasons why you should consider becoming a full stack developer -\n\n1. Building unique proprietary software products\nEvery company/developer follows their own tech stack while building a product (as every product has its own requirements and the best set of technology is chosen accordingly).\n\nThe use of such diverse technologies makes the code base of the project complex and this eventually creates a unique code for a particular product. Companies adopting full stack development will have this edge.",
        "status": "published",
        "date": "2021-11-20T14:23:37.117Z",
        "_id": "619904e92b88fabc684ab332",
        "createdAt": "2021-11-20T14:23:37.130Z",
        "updatedAt": "2021-11-20T14:23:37.130Z"
      },
      "type": "object",
      "properties": {
        "category_id": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "image_alt": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "author_id": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "_id": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      },
      "required": [
        "category_id",
        "image",
        "image_alt",
        "title",
        "author_id",
        "content",
        "status",
        "date",
        "_id",
        "createdAt",
        "updatedAt"
      ]
    },
    "UpdateBlog-UnAuthorized1": {
      "title": "UpdateBlog-UnAuthorized1",
      "example": {
        "status": 0,
        "message": "You are not Authorized!",
        "data": null
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      },
      "required": [
        "status",
        "message",
        "data"
      ]
    },
    "UpdateBlog-Authorized": {
      "title": "UpdateBlog-Authorized",
      "example": {
        "status": 1,
        "message": "Blog Updated Successfully.",
        "data": {
          "_id": "6197cf099d6a0fa882cb0e75",
          "category_id": "6197cd59e28a5eb64301b2c8",
          "image": "https://bloggify.s3.amazonaws.com/90dee1de-9a03-493a-90b2-c37cd4079306.jpeg",
          "image_alt": "90dee1de-9a03-493a-90b2-c37cd4079306",
          "title": "Why should you become a full stack developer?",
          "author_id": "61977e099fe7b649dae5e0c0",
          "content": "This diverse working knowledge gives developers the flexibility to tailor the tools needed for the project as per the project needs and create a state of the art products; since every technology has something different to offer - broadly in terms of syntaxes, compiling time (performance), compatibility, code reusability.\n\nOne fine example is that web developers with front-end knowledge can easily convert animations and graphics offered by software like Adobe Photoshop directly to front-end code. Thereby reducing the dependency on such third-party proprietary software. Limiting knowledge of basic styling tools will bind the developer to use such third-party software and thus burdening the project’s budget. Similar advantages apply to back-end scenarios. And thus knowing several essential tools will open up innovative approaches to projects.",
          "status": "published",
          "date": "2021-11-19T16:21:29.073Z",
          "createdAt": "2021-11-19T16:21:29.109Z",
          "updatedAt": "2021-11-20T14:26:12.847Z"
        }
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/Data2"
        }
      },
      "required": [
        "status",
        "message",
        "data"
      ]
    },
    "DeleteBlog-NotFound": {
      "title": "DeleteBlog-NotFound",
      "example": {
        "status": 0,
        "message": "Blog Not Found!",
        "data": null
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      },
      "required": [
        "status",
        "message",
        "data"
      ]
    },
    "DeleteBlog-NotAuthorized1": {
      "title": "DeleteBlog-NotAuthorized1",
      "example": {
        "status": 0,
        "message": "You are not Authorized!",
        "data": null
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      },
      "required": [
        "status",
        "message",
        "data"
      ]
    },
    "DeleteBlog-Success": {
      "title": "DeleteBlog-Success",
      "example": {
        "status": 1,
        "message": "Blog Deleted Successfully.",
        "data": {
          "_id": "619904e92b88fabc684ab332",
          "category_id": "6197cd59e28a5eb64301b2c8",
          "image": "https://bloggify.s3.amazonaws.com/90dee1de-9a03-493a-90b2-c37cd4079306.jpeg",
          "image_alt": "90dee1de-9a03-493a-90b2-c37cd4079306",
          "title": "Why should you become a full stack developer?",
          "author_id": "6199039f2b88fabc684ab32f",
          "content": "Let’s get this cleared first. Here are the top 3 reasons why you should consider becoming a full stack developer -\n\n1. Building unique proprietary software products\nEvery company/developer follows their own tech stack while building a product (as every product has its own requirements and the best set of technology is chosen accordingly).\n\nThe use of such diverse technologies makes the code base of the project complex and this eventually creates a unique code for a particular product. Companies adopting full stack development will have this edge.",
          "status": "published",
          "date": "2021-11-20T14:23:37.117Z",
          "createdAt": "2021-11-20T14:23:37.130Z",
          "updatedAt": "2021-11-20T14:23:37.130Z"
        }
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/Data2"
        }
      },
      "required": [
        "status",
        "message",
        "data"
      ]
    },
...
    "SearchUser-Success": {
      "title": "SearchUser-Success",
      "example": {
        "status": 1,
        "message": "User Found Success.",
        "data": {
          "email": "hkanani191@gmail.com",
          "usernmae": "harikanani",
          "full_name": "hari kanani"
        }
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/Data8"
        }
      },
      "required": [
        "status",
        "message",
        "data"
      ]
    },
    "Data8": {
      "title": "Data8",
      "example": {
        "email": "hkanani191@gmail.com",
        "usernmae": "harikanani",
        "full_name": "hari kanani"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "usernmae": {
          "type": "string"
        },
        "full_name": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "usernmae",
        "full_name"
      ]
    },
...
  },
...
}